requirements.txt
Django==2.0.7
django-rest-framework-mongoengine==3.3.1
django-bootstrap4==0.0.6
django.js==0.8.1
djongo==1.2.29
pymongo==3.7.1
pytz==2018.5
djangorestframework==3.8.2
mongoengine==0.15.3
